+++ src/core/index.test.ts
@@ -8,6 +8,7 @@ import {
 	IDeleteDnsRecordMessage,
 	ICreateDnsRecordContent,
 } from '../models/IDnsRecordMessage'
+import { ISeedBlock } from '../models/ISeedBlock'
 
 describe('Core module test suite', () => {
 	it('Hash Unit Test', async () => {
@@ -435,5 +436,140 @@ describe('Core module test suite', () => {
 		await core.deleteDnsRecord(signedDnsRecordMessage, 'http://localhost:3000')
 		expect(httpDelete).toHaveBeenCalled()
 		expect(true).toEqual(true)
+	})
+
+	it('Get Account Seed Block Unauthenticated Test', async () => {
+		// Initializations
+		const httpModule = new http()
+		const core = new Core(httpModule)
+
+		const accountSeedBlockMessage = {
+			content: {
+				accountPublicKey: 'accountPublicKey',
+			},
+		}
+		const hostingProviderAddresses = ['https://hosting.com']
+
+		// Mocking
+		const seedBlockExpectedResponse = {
+			version: 1,
+			address:
+				'0x0004003e0005007a001500c200b200740025007300ef00b1009900af00d4006800a800270074006800da009c008000e00069008700fc0059005a000e00f3001d00aa0029005500a400dc006000e9003900a5008f00140098002300dc00b600e000c1009c000e000900d5001800d800e3008f0061007e00470034003f007f0005007f',
+			public_key:
+				'0x000400e1009400a5007900620029009a00f1007f00a900bb005200b900b70045002a0034009b004a00b8008e004000c500e200860001000f00b0005100010016006100ed004c00050001002c00ad008700d6005700000025009a00e3005c00780099004b007e003400b700b5003d00cc00a3007100b200af005500d7005a0000009d',
+			update_id: 1,
+			signature:
+				'0x00be00d40099003b002f00b9009000dd0069002500a30012002e007700ef00620075008e006000250084003900780071000a002f006500df00eb002800d7006300bc00dc003e00a4002600fb00ca00b8000500cd00e600a600f20090004200cf007f00ca00c400b20062003d00f3009d00ee00b5002e00ea009d009b00670008',
+		}
+		const httpGet = jest.fn().mockImplementation(async () => ({
+			ok: true,
+			status: 200,
+			json: async () => seedBlockExpectedResponse,
+		}))
+		jest.spyOn(httpModule, 'post').mockImplementation(httpGet)
+
+		// Test
+		const seedBlockResponse = await core.getAccountSeedBlockUnauthenticated(
+			accountSeedBlockMessage,
+			hostingProviderAddresses,
+		)
+		expect(seedBlockResponse).toBeDefined()
+		if (seedBlockResponse) {
+			expect(seedBlockResponse === seedBlockExpectedResponse).toEqual(true)
+		}
+	})
+
+	it('Get Account Seed Block by Username Test', async () => {
+		// Initializations
+		const httpModule = new http()
+		const core = new Core(httpModule)
+
+		// Mocking
+		const getAccountSeedBlockUnauthenticatedResponse = {
+			version: 1,
+			address:
+				'0x0004003e0005007a001500c200b200740025007300ef00b1009900af00d4006800a800270074006800da009c008000e00069008700fc0059005a000e00f3001d00aa0029005500a400dc006000e9003900a5008f00140098002300dc00b600e000c1009c000e000900d5001800d800e3008f0061007e00470034003f007f0005007f',
+			public_key:
+				'0x000400e1009400a5007900620029009a00f1007f00a900bb005200b900b70045002a0034009b004a00b8008e004000c500e200860001000f00b0005100010016006100ed004c00050001002c00ad008700d6005700000025009a00e3005c00780099004b007e003400b700b5003d00cc00a3007100b200af005500d7005a0000009d',
+			update_id: 1,
+			signature:
+				'0x00be00d40099003b002f00b9009000dd0069002500a30012002e007700ef00620075008e006000250084003900780071000a002f006500df00eb002800d7006300bc00dc003e00a4002600fb00ca00b8000500cd00e600a600f20090004200cf007f00ca00c400b20062003d00f3009d00ee00b5002e00ea009d009b00670008',
+		}
+		const getAccountSeedBlockUnauthenticated = jest
+			.fn()
+			.mockImplementation(
+				async () => getAccountSeedBlockUnauthenticatedResponse,
+			)
+		jest
+			.spyOn(core, 'getAccountSeedBlockUnauthenticated')
+			.mockImplementation(getAccountSeedBlockUnauthenticated)
+
+		const getDnsRecordUnauthenticatedResponse: IDnsRecord = {
+			domainName: 'example.stw',
+			accountPublicKey:
+				'0x000400e1009400a5007900620029009a00f1007f00a900bb005200b900b70045002a0034009b004a00b8008e004000c500e200860001000f00b0005100010016006100ed004c00050001002c00ad008700d6005700000025009a00e3005c00780099004b007e003400b700b5003d00cc00a3007100b200af005500d7005a0000009d',
+			hostingProviderAddresses: ['https://hosting.com'],
+		}
+
+		const getDnsRecordUnauthenticated = jest
+			.fn()
+			.mockImplementation(async () => getDnsRecordUnauthenticatedResponse)
+		jest
+			.spyOn(core, 'getDnsRecordUnauthenticated')
+			.mockImplementation(getDnsRecordUnauthenticated)
+
+		// Test
+		const seedBlockResponse =
+			await core.getAccountSeedBlockByUsernameUnauthenticated('example.stw')
+		expect(seedBlockResponse).toBeDefined()
+		if (seedBlockResponse) {
+			expect(
+				seedBlockResponse === getAccountSeedBlockUnauthenticatedResponse,
+			).toEqual(true)
+		}
+	})
+
+	it('Create Account Seed Block Test', async () => {
+		// Initializations
+		const httpModule = new http()
+		const core = new Core(httpModule)
+
+		const accountkeyPair = await Core.generateSignatureKeyPair()
+		const accountPublicKey = accountkeyPair.publicKey
+		const accountPrivateKey = accountkeyPair.privateKey
+
+		const blockKeyPair = await Core.generateSignatureKeyPair()
+		const blockPublicKey = blockKeyPair.publicKey
+
+		const unsignedSeedBlock = await Core.generateSeedBlock(
+			accountPublicKey,
+			blockPublicKey,
+			1,
+			1,
+		)
+
+		const signedSeedBlock = await Core.signSeedBlock(
+			unsignedSeedBlock,
+			accountPrivateKey,
+		)
+
+		// Mocking
+		const createAccountSeedBlockResponse: ISeedBlock = signedSeedBlock
+		const httpPost = jest.fn().mockImplementation(async () => ({
+			ok: true,
+			status: 200,
+			json: async () => createAccountSeedBlockResponse,
+		}))
+		jest.spyOn(httpModule, 'post').mockImplementation(httpPost)
+
+		// Test
+		const seedBlockResponse = await core.createAccountSeedBlock(
+			signedSeedBlock,
+			['https://hosting.com'],
+		)
+		expect(seedBlockResponse).toBeDefined()
+		if (seedBlockResponse) {
+			expect(seedBlockResponse === createAccountSeedBlockResponse).toEqual(true)
+		}
 	})
 })